[user-managerd] Move guest's home to /home. Fixes JB#50808
There isn't any real reason to keep guest's home in /tmp since it lives
in a separate file system. Mount it under /home that for example tracker
works correctly. Note that the home directory is still different from
the /home/$USER scheme as it uses an underscore instead of a dash.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>