[user-managerd] Prevent removing system groups on failed user add. Fixes JB#52871
Change removeGroup to take gid instead of group name and move group
removal on failure inside LibUserHelper::addUser() as it knows better
how to deal with it.

Check in SailfishUserManager::addUser() that the group name is not
reserved. Also treat already existing home directory the same way
(increment the number at the end of username).

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>