[docs] Fix index file generation. JB#53870
The 'project' variable determines the index file base name. It needs to
match the output directory base name, i.e. what is assigned to
SAILFISH_QDOC.project in the .pro file. If these two do not match, other
modules are not able to refer to it with the 'depends' variable.

The 'url' variable is meant to point to the location where the
documentation is going to be published. It is used as a base to form
external links to THIS documentation module when it is refered by other
modules.